Finally added support for saving the current items in a crafting machine. Now, if I add a recipe that takes hundreds of items, not all progress will be lost when the player saves and loads again. This is good.
It should also make things easier when I add support for saving the positions and vectors of an item in motion, and fore storage containers etc. This is all great!

Finally, I noticed a small error where the pycrete crafting recipe was accepting two sand despite it being explicitly indicated as accepting one sand. This must be rectified ASAP. Likely a problem in the AddItemToLauncher function. GL boys